assignment4/requirements.txt
facenet_pytorch==2.6.0
matplotlib==3.9.0
numpy==1.26.4
pandas==2.2.2
Pillow==10.2.0
